tests: Cast drop-messages-interval type properly
The rtpjitterbuffer test drop_messages_interval uses a GstClockTime for
the message drop interval.  This property is defined as a guint.  On
systems with 64-bit time_t but 32-bit uint, this can cause the
g_object_set function to fail to read the arguments properly.

Fixes: #1656
Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/3580>

+++ subprojects/gst-plugins-good/tests/check/elements/rtpjitterbuffer.c
@@ -3207,7 +3207,7 @@ GST_START_TEST (test_drop_messages_interval)
   guint num_sent_msg = 0;
 
   g_object_set (h->element, "post-drop-messages", TRUE, NULL);
-  g_object_set (h->element, "drop-messages-interval", interval, NULL);
+  g_object_set (h->element, "drop-messages-interval", (guint) interval, NULL);
   next_seqnum = construct_deterministic_initial_state (h, latency_ms);
 
   /* Create a bus to get the drop message on */
